Fixed templating, long ago, added event stuff, and updated shaders to shader version 1.5 (OpenGL 3.2)

+++ ftec-core/src/engine/Engine.cpp
@@ -70,7 +70,7 @@ namespace ftec {
 			getWindow().poll();
 
 			double currentTime = glfwGetTime();
-			Time::deltaTime = previousTime - currentTime;
+			Time::deltaTime = (float) (previousTime - currentTime);
 			Time::runTime += Time::deltaTime;
 			Time::calculateSinCosTime();
