+++ ftec-core/src/graphics/Window.h
@@ -25,8 +25,8 @@ namespace ftec {
 
 		void setCursorMode(int mode);
 
-		inline float getWidth() { return m_Width; }
-		inline float getHeight() { return m_Height; }
+		inline float getWidth() { return (float)m_Width; }
+		inline float getHeight() { return (float)m_Height; }
 		
 		inline vec2f getMousePosition() const { return m_MousePosition; }
 		inline bool isCloseRequested() const { return m_CloseRequested; }
